ipa: rpi: sync: Add an implementation of the camera sync algorithm
In this implementation, the server sends data packets out onto the
network every 30 frames or so.

Clients listening for this packet will send frame length deltas back
to the pipeline handler to match the synchronisation of the server.

We can use wallclock timestamps so that the process will actually work
across networked Pis, but it does really on those wallclocks being
properly synchronised. We de-jitter our wallclock measurements (as
they're made in userspace) to match the more accurate kernel
SensorTimestamp value.

When the server's advertised "ready time" is reached, both client and
server will signal this through metadata back to their respective
controlling applications.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Arsen Mikovic <arsen.mikovic@raspberrypi.com>
Signed-off-by: Naushir Patuck <naush@raspberrypi.com>