libcamera: rkisp1: Create main buffer pool out of if(!isRaw)
The "useDewarper_" class variable is set to true during configure
only if the configuration is !isRaw.

The main path buffer pool creation can thus be moved out of the
if (!isRaw) block at allocateBuffers() time.

Cosmetic change that will make it easier to create a buffer pool
for the main and self paths unconditionally in future.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>